+++ FanControl/build.gradle.kts
@@ -33,7 +33,7 @@ kotlin {
 
                 implementation("org.jetbrains.compose.material3:material3-desktop:${extra["compose.version"] as String}")
                 implementation("org.json:json:20220924")
-                //implementation("com.github.wiiznokes:setting-sliding-windows:1.0.3")
+                //implementation("com.github.wiiznokes:setting-sliding-windows:1.0.4")
                 implementation("com.example:setting-sliding-windows-jvm:1.0.4")
             }
         }
